+++ test/lib/tree-functionalities.test.js
@@ -145,13 +145,21 @@ test("fullDownstreamChain kie-wb-distributions", async () => {
       project: "kiegroup/optaplanner",
       mapping: {
         dependencies: {
-          default: {
-            source: "7.x",
-            target: "master"
-          }
+          default: [
+            {
+              source: "7.x",
+              target: "master"
+            }
+          ]
         },
-        source: "master",
-        target: "7.x",
+        dependant: {
+          default: [
+            {
+              source: "master",
+              target: "7.x"
+            }
+          ]
+        },
         exclude: [
           "kiegroup/optaweb-employee-rostering",
           "kiegroup/optaweb-vehicle-routing"
@@ -222,13 +230,16 @@ test("fullDownstreamChain optaweb-employee-rostering", async () => {
       project: "kiegroup/optaplanner",
       mapping: {
         dependencies: {
-          default: {
-            source: "7.x",
-            target: "master"
-          }
+          default: [{ source: "7.x", target: "master" }]
         },
-        source: "master",
-        target: "7.x",
+        dependant: {
+          default: [
+            {
+              source: "master",
+              target: "7.x"
+            }
+          ]
+        },
         exclude: [
           "kiegroup/optaweb-employee-rostering",
           "kiegroup/optaweb-vehicle-routing"
@@ -239,13 +250,21 @@ test("fullDownstreamChain optaweb-employee-rostering", async () => {
       project: "kiegroup/optaweb-employee-rostering",
       mapping: {
         dependencies: {
-          default: {
-            source: "7.x",
-            target: "master"
-          }
+          default: [
+            {
+              source: "7.x",
+              target: "master"
+            }
+          ]
         },
-        source: "master",
-        target: "7.x",
+        dependant: {
+          default: [
+            {
+              source: "master",
+              target: "7.x"
+            }
+          ]
+        },
         exclude: ["kiegroup/optaweb-vehicle-routing", "kiegroup/optaplanner"]
       }
     }