+++ test/lib/util/chain-util.test.js
@@ -80,13 +80,21 @@ test("parentChainFromNode optaplanner", async () => {
       project: "kiegroup/optaplanner",
       mapping: {
         dependencies: {
-          default: {
-            source: "7.x",
-            target: "master"
-          }
+          default: [
+            {
+              source: "7.x",
+              target: "master"
+            }
+          ]
         },
-        source: "master",
-        target: "7.x",
+        dependant: {
+          default: [
+            {
+              source: "master",
+              target: "7.x"
+            }
+          ]
+        },
         exclude: [
           "kiegroup/optaweb-employee-rostering",
           "kiegroup/optaweb-vehicle-routing"
@@ -123,13 +131,21 @@ test("parentChainFromNode kie-wb-distributions", async () => {
       project: "kiegroup/optaplanner",
       mapping: {
         dependencies: {
-          default: {
-            source: "7.x",
-            target: "master"
-          }
+          default: [
+            {
+              source: "7.x",
+              target: "master"
+            }
+          ]
         },
-        source: "master",
-        target: "7.x",
+        dependant: {
+          default: [
+            {
+              source: "master",
+              target: "7.x"
+            }
+          ]
+        },
         exclude: [
           "kiegroup/optaweb-employee-rostering",
           "kiegroup/optaweb-vehicle-routing"